Update examples to use Parcel
