--- examples/README.md
@@ -1,9 +1,0 @@
-# Redux Promise Middleware Example
-
-The examples demonstrate using promise middleware to make async requests. There is also a simple example with just Redux. Both examples use [a mocked JSON API](https://github.com/typicode/json-server) with [Faker data](https://github.com/marak/Faker.js/).
-
-## Getting Started
-
-1. Install dependencies `npm install`
-2. Start the app: `npm start`
-3. Go to [localhost:8000](http://localhost:8000)
