+++ examples/catching-errors-with-middleware/.gitignore
@@ -1,0 +1,3 @@
+.cache
+dist
+node_modules
