+++ examples/catching-errors-with-middleware/index.html
@@ -1,0 +1,4 @@
+<!doctype html public>
+<body>
+<div id="mount"/>
+<script async src="./index.js"></script>
