+++ examples/catching-errors-with-middleware/index.js
@@ -1,0 +1,12 @@
+import React, { Component } from 'react';
+import { render } from 'react-dom';
+
+class App extends Component {
+  render() {
+    return (
+      <div />
+    );
+  }
+}
+
+render(<App />, document.querySelector('#mount'));
