+++ examples/catching-errors-with-middleware/middleware/error.js
@@ -1,0 +1,42 @@
+import isPromise from 'is-promise';
+import oneOfType from '../utils/oneOfType';
+
+export default function error() {
+  return next => action => {
+    const types = [
+      'GLOBAL_ERROR'
+    ];
+
+    // If not a promise, continue on
+    if (!isPromise(action.payload)) {
+      return next(action);
+    }
+
+    /*
+     * Because it iterates on an array for every async action, this
+     * oneOfType function could be expensive to call in production.
+     * Another solution would would be to include a property in `meta`
+     * and evaulate that property.
+     *
+     * if (action.meta.globalError === true) {
+     *   // handle error
+     * }
+     *
+     * The error middleware serves to dispatch the initial pending promise to
+     * the promise middleware, but adds a `catch`.
+     */
+    if (oneOfType(action.type, types)) {
+
+      // Dispatch initial pending promise, but catch any errors
+      return next(action).catch(error => {
+        if (process.env.NODE_ENV === 'development') {
+          console.warn(`${action.type} caught at middleware with reason: ${JSON.stringify(error.message)}.`);
+        }
+
+        return error;
+      });
+    }
+
+    return next(action);
+  };
+}
