+++ examples/catching-errors-with-middleware/middleware/oneOfType.js
