+++ examples/catching-errors-with-middleware/package.json
@@ -1,0 +1,19 @@
+{
+  "main": "index.js",
+  "scripts": {
+    "start": "`npm bin`/parcel index.html",
+    "test": ""
+  },
+  "dependencies": {
+    "is-promise": "^2.1.0",
+    "react": "^16.4.1",
+    "react-dom": "^16.4.1",
+    "react-redux": "^5.0.7",
+    "redux": "^4.0.0",
+    "redux-logger": "^3.0.6",
+    "redux-thunk": "^2.3.0"
+  },
+  "devDependencies": {
+    "parcel-bundler": "^1.9.3"
+  }
+}
