+++ examples/catching-errors-with-middleware/tests/middleware/error.spec.js
