--- examples/complex/actions/index.js
@@ -1,2 +1,0 @@
-export * as application from './application';
-export * as error from './error';
