--- examples/complex/constants/application.js
@@ -1,2 +1,0 @@
-export const APPLICATION_INITIALIZE_CHECK = 'APPLICATION_INITIALIZE_CHECK';
-export const APPLICATION_SERVER_CHECK = 'APPLICATION_SERVER_CHECK';
