--- examples/complex/constants/error.js
@@ -1,2 +1,0 @@
-export const GLOBAL_ERROR = 'GLOBAL_ERROR';
-export const LOCAL_ERROR = 'LOCAL_ERROR';
