--- examples/complex/index.html
@@ -1,5 +1,0 @@
-<!doctype html public>
-<base href="/complex"/>
-<body>
-<div id="mount"/>
-<script src="/complex.js"></script>
