--- examples/complex/reducers/application/index.js
@@ -1,5 +1,0 @@
-import { combineReducers } from 'redux';
-
-export default combineReducers({
-  support: require('./support').default
-});
