--- examples/complex/reducers/index.js
@@ -1,5 +1,0 @@
-import { combineReducers } from 'redux';
-
-export default combineReducers({
-  application: require('./application').default
-});
