--- examples/complex/routes.js
@@ -1,8 +1,0 @@
-import component from './components/app';
-import indexRoute from './routes/home';
-
-export default [{
-  path: '/',
-  component,
-  indexRoute
-}];
