--- examples/complex/routes/home/container.js
@@ -1,14 +1,0 @@
-import { connect } from 'react-redux';
-import * as errorActions from '../../actions/error';
-
-export default connect(
-  state => ({}),
-  dispatch => ({
-    throwGlobalError: message => {
-      dispatch(errorActions.rejectPromiseWithGlobalError(message))
-    },
-    throwLocalError: message => {
-      dispatch(errorActions.rejectPromiseWithLocalError(message))
-    }
-  })
-)(require('./component').default);
