--- examples/complex/routes/home/index.js
@@ -1,3 +1,0 @@
-export default {
-  component: require('./container').default,
-};
