--- examples/complex/utils/isPromise.js
@@ -1,7 +1,0 @@
-export default function isPromise(value) {
-  if (value !== null && typeof value === 'object') {
-    return value && typeof value.then === 'function';
-  }
-
-  return false;
-}
