--- examples/promise-all/index.html
@@ -1,6 +1,0 @@
-<!doctype html public>
-<body>
-<button id="load">Get All</button>
-<hr />
-<div id="mount"/>
-<script src="/promise-all.js"></script>
