--- examples/simple/README.md
@@ -1,17 +1,0 @@
-# Simple Example
-
-This example demonstrates how to use the middleware at a basic level with just Redux and no React.
-
-## Getting Started
-
-1. Navigate to the examples: `cd examples`
-2. Install dependencies `npm install`
-2. Start the app: `npm start`
-4. Go to [localhost:8000/simple](http://localhost:8000/simple)
-
-## Logging
-
-If you open the developer console in your web browser, the actions are logged with [Redux Logger](https://github.com/evgenyrodionov/redux-logger) for you to see the state change.
-
----
-Copyright (c) 2018 Patrick Burtchaell. [Code licensed with the MIT License (MIT)](/LICENSE). [Documentation licensed with the CC BY-NC License](/docs/LICENSE).
