--- examples/simple/index.html
@@ -1,6 +1,0 @@
-<!doctype html public>
-<body>
-<button id="load">Get Post</button>
-<hr />
-<div id="mount"/>
-<script src="/simple.js"></script>
