+++ examples/using-promise-all/store.js
@@ -1,0 +1,28 @@
+import { createStore, applyMiddleware } from 'redux';
+import promiseMiddleware from '../../../src/index';
+import thunkMiddleware from 'redux-thunk';
+import logger from 'redux-logger';
+
+const reducer = (state = {}, action) => {
+  switch (action.type) {
+    case 'GET_ALL_PENDING':
+      return {
+        isPending: true,
+      };
+
+    case 'GET_ALL_FULFILLED':
+      return {
+        isPending: false,
+      };
+
+    default: return state;
+  }
+};
+
+const store = createStore(reducer, {}, applyMiddleware(
+  thunkMiddleware,
+  promiseMiddleware(),
+  logger,
+));
+
+export default store;
