+++ examples/using-promise-middleware/index.html
@@ -1,0 +1,6 @@
+<!doctype html public>
+<body>
+<button id="load">Get Post</button>
+<hr />
+<div id="mount"/>
+<script async src="./index.js"></script>
