+++ examples/using-promise-middleware/index.js
