+++ examples/using-promise-middleware/package.json
@@ -1,0 +1,14 @@
+{
+  "main": "index.js",
+  "scripts": {
+    "start": "`npm bin`/parcel index.html"
+  },
+  "dependencies": {
+    "redux": "^4.0.0",
+    "redux-logger": "^3.0.6",
+    "redux-thunk": "^2.3.0"
+  },
+  "devDependencies": {
+    "parcel-bundler": "^1.9.3"
+  }
+}
